📝 feat. Card 컴포 폴더명 및 경로변경
