+++ src/components/mainPage/NewCarousel.tsx
@@ -1,94 +1,94 @@
-import styled from "styled-components";
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
+import styled from 'styled-components'
+import Slider from 'react-slick'
+import 'slick-carousel/slick/slick.css'
+import 'slick-carousel/slick/slick-theme.css'
 
-import "../../style/carousel.css";
-import Card from "../Card";
+import '../../style/carousel.css'
+import Card from '../card/Card'
 
 const NewCarousel = () => {
   // carousel option
   const setting = {
     autoplaySpeed: 3000,
     autoplay: true,
-    dotsClass: "slick-dots",
+    dotsClass: 'slick-dots',
     dots: true,
     slidesToShow: 3,
     slidesToScroll: 3,
-  };
+  }
 
   return (
     <MainContainer>
       <Slider {...setting}>
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
         <Card
           p_Thumbnail={
-            "https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg"
+            'https://cdn.imweb.me/thumbnail/20220428/8a3716a1e9c28.jpg'
           }
-          p_Brand={"지구샵"}
-          p_Name={"키친클로스(체크)"}
+          p_Brand={'지구샵'}
+          p_Name={'키친클로스(체크)'}
         />
       </Slider>
     </MainContainer>
-  );
-};
+  )
+}
 
-export default NewCarousel;
+export default NewCarousel
 
 const MainContainer = styled.div`
   width: 55%;
@@ -96,4 +96,4 @@ const MainContainer = styled.div`
     width: 100%;
     margin: 0 auto;
   }
-`;
+`
